net: mvpp2: rx-queue structure pack to cache line
Align RX queue structure onto L1-cache-line and place frequently
used fields on top of the structure.
Compact the field "logic_rxq" from 4bytes onto 1byte for better
cache utilization.

Reviewed-on: http://vgitil04.il.marvell.com:8080/60520
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]

+++ drivers/net/ethernet/marvell/mvpp2/mvpp2_main.c
@@ -5187,7 +5187,7 @@ static int mvpp2_port_init(struct mvpp2_port *port)
 		/* Map this Rx queue to a physical queue */
 		rxq->id = port->first_rxq + queue;
 		rxq->port = port->id;
-		rxq->logic_rxq = queue;
+		rxq->logic_rxq = (u8)queue;
 
 		port->rxqs[queue] = rxq;
 	}
